Refactor email sending jobs to extract common functionality into concerns rather than have base classes that cannot be instaniated directly
Move all shared functionality for emailing jobs into concerns out of concrete job classes.

Use usec accuracy when comparing versions of petitions.
