+++ app/jobs/deliver_debate_outcome_email_job.rb
@@ -1,0 +1,8 @@
+class DeliverDebateOutcomeEmailJob < ActiveJob::Base
+  include EmailDelivery
+
+  def create_email
+    mailer.notify_signer_of_debate_outcome signature.petition, signature
+  end
+
+end
