+++ app/jobs/deliver_debate_scheduled_email_job.rb
@@ -1,0 +1,8 @@
+class DeliverDebateScheduledEmailJob < ActiveJob::Base
+  include EmailDelivery
+
+  def create_email
+    mailer.notify_signer_of_debate_scheduled signature.petition, signature
+  end
+
+end
