+++ app/jobs/deliver_threshold_response_email_job.rb
@@ -1,0 +1,8 @@
+class DeliverThresholdResponseEmailJob < ActiveJob::Base
+  include EmailDelivery
+
+  def create_email
+    mailer.notify_signer_of_threshold_response signature.petition, signature
+  end
+
+end
