+++ app/jobs/email_debate_scheduled_job.rb
@@ -1,6 +1,8 @@
-class EmailDebateScheduledJob < EmailPetitionSignatoriesJob
+class EmailDebateScheduledJob < ActiveJob::Base
+  include EmailAllPetitionSignatories
+
   def email_delivery_job_class
-    EmailDeliveryJobs::DebateScheduled
+    DeliverDebateScheduledEmailJob
   end
 
   def timestamp_name
