--- app/jobs/email_delivery_jobs/debate_scheduled.rb
@@ -1,9 +1,0 @@
-module EmailDeliveryJobs
-  class DebateScheduled < Base
-
-    def create_email
-      mailer.notify_signer_of_debate_scheduled signature.petition, signature
-    end
-
-  end
-end
