--- app/jobs/email_delivery_jobs/threshold_response.rb
@@ -1,9 +1,0 @@
-module EmailDeliveryJobs
-  class ThresholdResponse < Base
-
-    def create_email
-      mailer.notify_signer_of_threshold_response signature.petition, signature
-    end
-
-  end
-end
