+++ app/jobs/email_threshold_response_job.rb
@@ -1,19 +1,11 @@
-class EmailThresholdResponseJob < EmailPetitionSignatories::Job
-  def self.run_later_tonight(petition, requested_at = Time.current)
-    petition.set_email_requested_at_for('government_response', to: requested_at)
-    super(petition, petition.get_email_requested_at_for('government_response'))
-  end
+class EmailThresholdResponseJob < ActiveJob::Base
+  include EmailAllPetitionSignatories
 
-  def perform(petition, requested_at_string, mailer = PetitionMailer.name, logger = nil)
-    @mailer = mailer.constantize
-    worker(petition, requested_at_string, logger).do_work!
+  def email_delivery_job_class
+    DeliverThresholdResponseEmailJob
   end
 
   def timestamp_name
     'government_response'
-  end
-
-  def create_email(petition, signature)
-    @mailer.notify_signer_of_threshold_response(petition, signature)
   end
 end
